AGI/requirements.txt
torch>=2.0.0
transformers>=4.35.0
pillow>=10.0.0
pydantic>=2.0
numpy
structlog
pytest
pytest-asyncio
pyyaml
